fix: Wire circuit breaker, fix cost truncation, make pricing configurable (#12 #11 #10)
- Add CircuitBreakerTomlConfig and PricingTomlConfig structs
- Add circuit_breaker and pricing fields to TomlConfig and Config
- Wire circuit breaker into pipeline via with_retry() when configured
- Fix cost truncation: return f64 cents instead of u64 for sub-cent precision
- Add default_pricing() with hardcoded fallbacks; allow .reviewer.toml overrides
- Update ReviewMetrics.estimated_cost_cents to f64
- Update all tests for new signatures and precision

+++ src/output.rs
@@ -28,7 +28,7 @@ pub struct ReviewMetrics {
     /// API latency in seconds.
     pub latency_secs: f64,
     /// Estimated cost in cents (USD).
-    pub estimated_cost_cents: u64,
+    pub estimated_cost_cents: f64,
     /// Diff size in lines.
     pub diff_lines: usize,
     /// Parsed verdict string.
@@ -221,7 +221,7 @@ mod tests {
             estimated_tokens_in: 4230,
             estimated_tokens_out: 892,
             latency_secs: 8.4,
-            estimated_cost_cents: 3,
+            estimated_cost_cents: 3.0,
             diff_lines: 150,
             verdict: "POSITIVE".to_string(),
             state: "APPROVE".to_string(),
